problemset-76282/doobestan/requirements.txt
asgiref==3.5.2
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
h11==0.13.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
uvicorn==0.17.6
